gumroad/polymarket-starter-kit/requirements.txt
polymarket-apis>=0.1.0
python-dotenv>=1.0.0
